Fix integer overflow in nccl thunk (#96)

+++ tensorflow/compiler/xla/service/gpu/nccl_all_to_all_thunk.cc
@@ -98,7 +98,7 @@ Status NcclAllToAllThunk::RunNcclCollective(const ExecuteParams& params,
       TF_ASSIGN_OR_RETURN(auto dtype_and_multiplier,
                           ToNcclDataTypeAndCountMultiplier(element_type));
       ncclDataType_t dtype = dtype_and_multiplier.first;
-      int element_count = buffer.element_count * dtype_and_multiplier.second;
+      int64_t element_count = buffer.element_count * dtype_and_multiplier.second;
 
       TF_RET_CHECK(element_count % num_participants == 0)
           << "Buffer was not an exact multiple of the number of participants.";
@@ -132,7 +132,7 @@ Status NcclAllToAllThunk::RunNcclCollective(const ExecuteParams& params,
       TF_ASSIGN_OR_RETURN(auto dtype_and_multiplier,
                           ToNcclDataTypeAndCountMultiplier(element_type));
       ncclDataType_t dtype = dtype_and_multiplier.first;
-      int element_count = buffer.element_count * dtype_and_multiplier.second;
+      int64_t element_count = buffer.element_count * dtype_and_multiplier.second;
 
       XLA_CUDA_RETURN_IF_ERROR(ncclSend(send_buffer, element_count, dtype,
                                         /*rank=*/i, comm, *cu_stream));
